+++ spring-boot/src/main/java/com/example/reviewapp/entity/Review.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public class Review {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     
     private String courseId;
     private String className;
@@ -24,12 +24,11 @@ public class Review {
     @JoinColumn(name = "student_id")
     private Student student;
     
-    // Getters and Setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
     
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
