+++ spring-boot/src/main/java/com/example/reviewapp/entity/Student.java
@@ -9,7 +9,7 @@ public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "student_id")
-    private Long id;
+    private Integer id;
     
     @OneToOne
     @JoinColumn(name = "user_id")
@@ -21,12 +21,11 @@ public class Student {
     @OneToMany(mappedBy = "student")
     private List<Review> reviews;
     
-    // Getters and Setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
     
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
